src/server/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiohttp-cors==0.7.0
aiosignal==1.3.2
attrs==25.2.0
bidict==0.23.1
blinker==1.9.0
bson==0.5.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
eventlet==0.39.1
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.3.0
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.32.3
setuptools==75.8.0
simple-websocket==1.1.0
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
yarl==1.18.3
